+++ src/org/epics/pvaccess/easyPVA/EasyMultiChannel.java
@@ -3,21 +3,28 @@
  */
 package org.epics.pvaccess.easyPVA;
 
-import org.epics.pvaccess.client.ChannelRequester;
+import org.epics.pvdata.pv.MessageType;
 import org.epics.pvdata.pv.PVStructure;
 import org.epics.pvdata.pv.Status;
+import org.epics.pvdata.pv.Union;
 
 /**
  * An easy to use interface to get/put data from/to multiple channels.
  * @author mrk
  *
  */
-public interface EasyMultiChannel extends ChannelRequester{
+public interface EasyMultiChannel {
     /**
-     * Get the name of the channelProvider.
+     * Get the requester name.
      * @return The name.
      */
-    String getProviderName();
+    public String getRequesterName();
+    /**
+     * new message.
+     * @param message The message string.
+     * @param messageType The message type.
+     */
+    public void message(String message, MessageType messageType);
     /**
      * Get the channelNames.
      * @return the names.
@@ -34,6 +41,13 @@ public interface EasyMultiChannel extends ChannelRequester{
      */
     boolean connect(double timeout);
     /**
+     * Calls issueConnect and then waitConnect.
+     * @param timeout timeOut for waitConnect
+     * @param minConnect minConnect for waitConnect
+     * @return (false,true) means (did not, did) connect.
+     */
+    boolean connect(double timeout,int minConnect);
+    /**
      * Connect to all channels.
      */
     void issueConnect();
@@ -45,6 +59,14 @@ public interface EasyMultiChannel extends ChannelRequester{
      */
     boolean waitConnect(double timeout);
     /**
+     * Wait until minConnect channels are connected or until timeout and no more channels connect.
+     * @param timeout The time to wait for connections.
+     * @param minConnect The minimum number of channels that must connect.
+     * When a timeout occurs a new time out will start if any channels connected since the last timeout.
+     * @return (false,true) of all channels (did not, did) connect.
+     */
+    boolean waitConnect(double timeout,int minConnect);
+    /**
      * Are all channels connected?
      * @return if all are connected.
      */
@@ -54,6 +76,13 @@ public interface EasyMultiChannel extends ChannelRequester{
      * @return The state of each channel.
      */
     boolean[] isConnected();
+    /**
+     * Create an EasyMultiData.
+     * @param pvRequest The pvRequest for each channel.
+     * @param union The union for each channel.
+     * @return The interface.
+     */
+    EasyMultiData createEasyMultiData(PVStructure pvRequest,Union union);
     /**
      * create a multiChannelGet that presents data as a NTMultiChannel.
      * calls the next method with request = "field(value,alarm,timeStamp)"
@@ -105,6 +134,46 @@ public interface EasyMultiChannel extends ChannelRequester{
      * @return EasyMultiPut or null if invalid request.
      */
     EasyMultiPut createPut(boolean doubleOnly);
+    /**
+     * Call the next method with request =  "field(value,alarm,timeStamp)" 
+     * @return The interface.
+     */
+    EasyMultiMonitor createMonitor();
+    /**
+     * First call createRequest as implemented by pvDataJava and then calls the next method.
+     * @param request The request as described in package org.epics.pvdata.copy
+     * @return The interface.
+     */
+    EasyMultiMonitor createMonitor(String request);
+    /**
+     * Creates an EasyMultiMonitor.
+     * The pvRequest is used to create the monitor for each channel.
+     * @param pvRequest The syntax of pvRequest is described in package org.epics.pvdata.copy.
+     * @return The interface.
+     */
+    EasyMultiMonitor createMonitor(PVStructure pvRequest);
+    /**
+     * Call the next method with request =  "field(value,alarm,timeStamp)" 
+     * @param doubleOnly true if data must be presented as a double[].
+     * @return The interface.
+     */
+    EasyMultiMonitor createMonitor(boolean doubleOnly);
+    /**
+     * First call createRequest as implemented by pvDataJava and then calls the next method.
+     * @param doubleOnly true if data must be presented as a double[].
+     * @param request The request as described in package org.epics.pvdata.copy
+     * @return The interface.
+     */
+    EasyMultiMonitor createMonitor(boolean doubleOnly,String request);
+    /**
+     * Creates an EasyMultiMonitor.
+     * The pvRequest is used to create the monitor for each channel.
+     * @param doubleOnly true if data must be presented as a double[].
+     * @param pvRequest The syntax of pvRequest is described in package org.epics.pvdata.copy.
+     * @return The interface.
+     */
+    EasyMultiMonitor createMonitor(boolean doubleOnly,PVStructure pvRequest);
+   
     /**
      * Set a new status value. The new value will replace the current status. The initial status is statusOK.
      * @param status The new status.